Move library modules to lib folder
